Implement Acess Port

+++ jocd/src/br/org/certi/jocd/dapaccess/Transfer.java
@@ -25,7 +25,7 @@ public class Transfer {
 
   private Exception error;
   private int sizeBytes = 0;
-  private int[] result;
+  private long[] result;
   private DapAccessCmsisDap dapLink;
   private byte dapIndex;
   private int transferCount;
@@ -64,7 +64,7 @@ public class Transfer {
   public void addResponse(byte[] data) {
     assert data.length == this.sizeBytes;
     int resultSize = this.sizeBytes / 4;
-    int[] result = new int[resultSize];
+    long[] result = new long[resultSize];
     int count = 0;
     for (int i = 0; i < this.sizeBytes; i += 4) {
       int word = (((data[0 + i] << 0) | (data[1 + i] << 8) | (data[2 + i] << 16) | (
@@ -86,7 +86,7 @@ public class Transfer {
   /*
    * Get the result of this transfer.
    */
-  public int[] getResult() throws Exception {
+  public long[] getResult() throws Exception {
     while (this.result == null) {
       if (this.dapLink.getCommandsToRead().size() > 0) {
         this.dapLink.readPacket();
